chore: move not common parts of General settings to specific places
