+++ src/components/pages/TestSuites/TestSuiteDetails/TestSuiteSettings/TestSuiteSettings.tsx
@@ -4,14 +4,14 @@ import {SettingsLayout} from '@molecules';
 import {SettingsLayoutProps} from '@molecules/SettingsLayout/SettingsLayout';
 
 import SettingsDefinition from '@organisms/EntityDetails/EntityDetailsContent/Settings/SettingsDefinition/SettingsDefinition';
-import General from '@organisms/EntityDetails/EntityDetailsContent/Settings/SettingsGeneral';
 import SettingsScheduling from '@organisms/EntityDetails/EntityDetailsContent/Settings/SettingsScheduling';
 import SettingsVariables from '@organisms/EntityDetails/EntityDetailsContent/Settings/SettingsVariables';
 
+import SettingsGeneral from './SettingsGeneral';
 import SettingsTests from './SettingsTests';
 
 const tabs = [
-  {id: 'general', label: 'General', children: <General />},
+  {id: 'general', label: 'General', children: <SettingsGeneral />},
   {id: 'tests', label: 'Tests', children: <SettingsTests />},
   {id: 'variables', label: 'Variables & Secrets', children: <SettingsVariables />},
   {id: 'scheduling', label: 'Scheduling', children: <SettingsScheduling />},
