+++ src/components/pages/Tests/TestDetails/TestSettings/SettingsGeneral/index.ts
@@ -1,0 +1,1 @@
+export {default} from './SettingsGeneral';
