+++ src/components/pages/Tests/TestDetails/TestSettings/TestSettings.tsx
@@ -4,15 +4,15 @@ import {SettingsLayout} from '@molecules';
 import {SettingsLayoutProps} from '@molecules/SettingsLayout/SettingsLayout';
 
 import SettingsDefinition from '@organisms/EntityDetails/EntityDetailsContent/Settings/SettingsDefinition/SettingsDefinition';
-import General from '@organisms/EntityDetails/EntityDetailsContent/Settings/SettingsGeneral';
 import SettingsScheduling from '@organisms/EntityDetails/EntityDetailsContent/Settings/SettingsScheduling';
 import SettingsVariables from '@organisms/EntityDetails/EntityDetailsContent/Settings/SettingsVariables';
 
 import SettingsExecution from './SettingsExecution';
+import SettingsGeneral from './SettingsGeneral';
 import SettingsTest from './SettingsTest';
 
 const tabs = [
-  {id: 'general', label: 'General', children: <General />},
+  {id: 'general', label: 'General', children: <SettingsGeneral />},
   {id: 'test', label: 'Test', children: <SettingsTest />},
   {id: 'execution', label: 'Execution', children: <SettingsExecution />},
   {id: 'variables', label: 'Variables & Secrets', children: <SettingsVariables />},
